+++ Src/ImplicitNullability.Plugin.Tests/ZoneMarkerAndSetUpFixture.cs
@@ -2,15 +2,19 @@
 using JetBrains.Application.BuildScript.Application.Zones;
 using JetBrains.ReSharper.TestFramework;
 using JetBrains.TestFramework;
-using JetBrains.TestFramework.Application.Zones;
 using NUnit.Framework;
+#if RS20181
+using ITestsEnvZone = JetBrains.TestFramework.Application.Zones.ITestsZone;
+#else
+using JetBrains.TestFramework.Application.Zones;
+#endif
 
 [assembly: RequiresSTA]
 
 namespace ImplicitNullability.Plugin.Tests
 {
     [ZoneDefinition]
-    public interface IImplicitNullabilityTestEnvironmentZone : ITestsZone, IRequire<PsiFeatureTestZone>, IRequire<IImplicitNullabilityZone>
+    public interface IImplicitNullabilityTestEnvironmentZone : ITestsEnvZone, IRequire<PsiFeatureTestZone>, IRequire<IImplicitNullabilityZone>
     {
     }