+++ Src/ImplicitNullability.Samples.CodeWithIN/Highlighting/IncorrectNullableAttributeUsageAnalyzer/OtherWarningsSample.cs
@@ -9,6 +9,8 @@ namespace ImplicitNullability.Samples.CodeWithIN.Highlighting.IncorrectNullableA
         public interface IInterface
         {
             void ExplicitCanBeNull([CanBeNull] string a);
+
+            string Property { get; set; }
         }
 
         public class Implementation : IInterface
@@ -30,6 +32,11 @@ namespace ImplicitNullability.Samples.CodeWithIN.Highlighting.IncorrectNullableA
 
             [NotNull, CanBeNull]
             public string MultipleAnnotations /*Expect:MultipleNullableAttributesUsage*/() => "";
+
+            [NotNull] /*Expect:NotObservableAnnotationRedundancy[RS > 20181 && !Implicit]*/
+            // In case of enabled IN, the (suppressed) AnnotationRedundancyInHierarchy hides the NotObservableAnnotationRedundancy
+            // ReSharper disable once NotNullMemberIsNotInitialized - it's not about this warning
+            string IInterface.Property { get; set; }
         }
     }
 }